+++ comics/comics/spikedmath.py
@@ -17,6 +17,9 @@ class Crawler(CrawlerBase):
         feed = self.parse_feed('http://feeds.feedburner.com/SpikedMath')
         for entry in feed.for_date(pub_date):
             page = self.parse_page(entry.link)
-            url = page.src('div.asset-body img')
-            title = entry.title
-            return CrawlerImage(url, title)
+            result = []
+            for url in page.src('div.asset-body img', allow_multiple=True):
+                result.append(CrawlerImage(url))
+            if result:
+                result[0].title = entry.title
+            return result